Fix/detail (#143)
* feat: transDateTime 함수 수정

* fix: 서버 응답과 다른 데이터명으로 인한 오류 수정
