+++ client/src/components/detail/CalloutBox.tsx
@@ -1,6 +1,6 @@
 import Button from "../Buttons";
 import styled from "styled-components";
-import { transDateTime } from "../main/transDateTime";
+import { transDateTime } from "../../util/transDateTime";
 
 const Container = styled.aside`
   width: 340px;
@@ -41,7 +41,7 @@ export default function CalloutBox({
     <Container>
       <div className="time">
         <span className="title">
-          {data.type === "buy" ? "업무 시간" : "희망 시간"}
+          {data.contentType === "BUY" ? "업무 시간" : "희망 시간"}
         </span>
         <ul>
           {transDateTime(data.workTimes).map((item, idx) => (
@@ -51,13 +51,13 @@ export default function CalloutBox({
       </div>
       <div>
         <span className="title">
-          {data.type === "buy" ? "장소" : "희망 장소"}
+          {data.contentType === "BUY" ? "장소" : "희망 장소"}
         </span>
         <span>{data.cityName}</span>
       </div>
       <div>
         <span className="title">
-          {data.type === "buy" ? "보수" : "희망 보수"}
+          {data.contentType === "BUY" ? "보수" : "희망 보수"}
         </span>
         <span>
           {data.price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}원
