+++ client/src/components/hunting/HuntingArticle.tsx
@@ -128,7 +128,9 @@ const HuntingArticle: React.FC = () => {
                   <span className="sub-title">작성자</span> {card.nickName}
                 </CardWriter>
                 <CardPay>
-                  <span className="sub-title">보수</span> {card.price}
+                  <span className="sub-title">보수</span>{" "}
+                  {card.price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",")}
+                  원
                 </CardPay>
                 {card.workTimes && (
                   <>
